Move tests to model folder
