+++ tests/models/layoutlmv3/test_tokenization_layoutlmv3.py
@@ -26,7 +26,7 @@ from transformers import AddedToken, LayoutLMv3TokenizerFast, SpecialTokensMixin
 from transformers.models.layoutlmv3.tokenization_layoutlmv3 import VOCAB_FILES_NAMES, LayoutLMv3Tokenizer
 from transformers.testing_utils import is_pt_tf_cross_test, require_pandas, require_tokenizers, require_torch, slow
 
-from ..test_tokenization_common import SMALL_TRAINING_CORPUS, TokenizerTesterMixin, merge_model_tokenizer_mappings
+from ...test_tokenization_common import SMALL_TRAINING_CORPUS, TokenizerTesterMixin, merge_model_tokenizer_mappings
 
 
 @require_tokenizers
